Fixed some goroutine leaks in tests
