+++ raftstore/prophet_adapter_test.go
@@ -1,3 +1,16 @@
+// Copyright 2020 MatrixOrigin.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package raftstore
 
 import (
@@ -12,6 +25,8 @@ import (
 	"github.com/matrixorigin/matrixcube/util/task"
 	"github.com/stretchr/testify/assert"
 )
+
+// FIXME: add leaktest checks
 
 func TestResourceAdapter(t *testing.T) {
 	ma := NewResourceAdapterWithShard(Shard{}).(*resourceAdapter)
