+++ raftstore/testutil.go
@@ -1076,6 +1076,14 @@ func (c *testRaftCluster) Stop() {
 	}
 }
 
+func (c *testRaftCluster) closeLogDBKVStorage() {
+	for _, s := range c.stores {
+		if s != nil && s.kvStorage != nil {
+			s.kvStorage.Close()
+		}
+	}
+}
+
 func (c *testRaftCluster) closeStorage() {
 	for _, s := range c.dataStorages {
 		if s != nil {
@@ -1401,5 +1409,8 @@ func newTestStore(t *testing.T) (*store, func()) {
 		_, err := ds.GetInitialStates()
 		assert.NoError(t, err)
 	}
-	return c.GetStore(0).(*store), func() { c.closeStorage() }
+	return c.GetStore(0).(*store), func() {
+		c.closeStorage()
+		c.closeLogDBKVStorage()
+	}
 }
